requirements.txt
absl-py==0.13.0
anyio==3.3.0
appdirs==1.4.3
argon2-cffi==20.1.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.2.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==4.0.0
blinker==1.4
cachetools==4.2.2
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
checkbox-support==0.22
clang==5.0
click==8.0.0
colorama==0.4.3
crcmod==1.7
cryptography==2.8
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
entrypoints==0.3
executing==0.8.2
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.5.2
fasteners==0.14.1
feedparser==5.2.1
filelock==3.0.12
Flask==2.0.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.39.0
guacamole==0.9.2
h5py==3.1.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
ImageHash==4.2.1
imageio==2.9.0
importlib-metadata==1.5.0
iniconfig==1.1.1
ipykernel==6.2.0
ipyplot==1.1.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.0
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.7.1
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.3.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
librosa==0.8.1
littleutils==0.2.2
llvmlite==0.37.0
lockfile==0.12.2
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mtcnn==0.1.1
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
nlpaug==1.1.7
notebook==6.4.3
numba==0.54.0
numpy==1.19.5
nvidia-ml-py3==7.352.0
nvidia-smi==0.1.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.3.56
opt-einsum==3.3.0
overrides==6.1.0
packaging==21.0
padme==1.1.1
pandas==1.3.2
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
pbr==5.6.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.3.1
plainbox==0.25
pluggy==1.0.0
pooch==1.5.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.5.1
ptyprocess==0.7.0
py==1.10.0
pyaml==21.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydub==0.25.1
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.18.0
PySocks==1.6.8
pytest==6.2.5
python-dateutil==2.7.3
python-speech-features==0.6
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resampy==0.2.2
rsa==4.7.2
scenedetect==0.5.6.1
scikit-learn==0.24.2
scikit-optimize==0.8.1
scikit-plot==0.3.7
scipy==1.7.1
SecretStorage==2.3.1
Send2Trash==1.8.0
shortuuid==1.0.1
simplejson==3.16.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sorcery==0.2.1
SoundFile==0.10.3.post1
soupsieve==1.9.5
ssh-import-id==5.10
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
testresources==2.0.1
threadpoolctl==2.2.0
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
traitlets==5.0.5
typing-extensions==3.7.4.3
typing-utils==0.0.3
urllib3==1.25.8
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
XlsxWriter==1.1.2
zipp==1.0.0

dlib~=19.22.1
mediapipe~=0.8.9
torchsummary~=1.5.1